+++ CHANGELOG.md
@@ -15,6 +15,7 @@
  * Add `Iban` constraint.
  * Add `Isbn` constraint.
  * Refactor `ZipCode` to get new countries option. See related PR: [ronanguilloux/IsoCodes#38](https://github.com/ronanguilloux/IsoCodes/pull/38)
+ * Make Symfony and Silex integration internal on this project.
 
 * 1.0.4 (2015-06-23)